refactor(FileList): move zarr logic out to hook
